Rename utils now that its functionality is more specific. Fix imports.
